Move all utils to single package
