+++ app/src/main/java/com/stochanskyi/librariesdemo/presentaiton/utils/PendingIntentUtils.kt
@@ -1,4 +1,4 @@
-package com.stochanskyi.librariesdemo.utils
+package com.stochanskyi.librariesdemo.presentaiton.utils
 
 import android.app.PendingIntent
 import android.os.Build
